backend/requirements.txt
fastapi
uvicorn
openai
pydantic
google-cloud-firestore
